fix: header, image - backspace deletes blocks only with no repeat key

+++ front/src/utils/editorjs/simple-image-plugin/SimpleImage/useSimpleImage.js
@@ -1,9 +1,19 @@
 import { useCallback, useEffect, useRef } from 'react';
 
+import { moveCaretToEnd } from '@sb-ui/utils/editorjs/toolsHelper';
 import { setPropsInTool } from '@sb-ui/utils/editorjs/utils';
 
 const isEmpty = (event, input) =>
   event.code === 'Backspace' && input?.innerText?.trim?.()?.length === 0;
+
+const stopRepeating = (event) => {
+  if (event.repeat) {
+    event.preventDefault();
+    event.stopPropagation();
+    return true;
+  }
+  return false;
+};
 
 export const useSimpleImage = ({ tool, loaded, error, src, setSrc }) => {
   const linkInputRef = useRef(null);
@@ -12,6 +22,9 @@ export const useSimpleImage = ({ tool, loaded, error, src, setSrc }) => {
   const handleInputLinkKeyDown = useCallback(
     (event) => {
       if (isEmpty(event, linkInputRef.current)) {
+        if (stopRepeating(event)) {
+          return;
+        }
         tool.api.blocks.delete();
       }
     },
@@ -20,7 +33,12 @@ export const useSimpleImage = ({ tool, loaded, error, src, setSrc }) => {
 
   const handleInputCaptionKeyDown = useCallback((event) => {
     if (isEmpty(event, captionInputRef.current)) {
+      if (stopRepeating(event)) {
+        return;
+      }
       linkInputRef.current.focus();
+      moveCaretToEnd(linkInputRef.current);
+      event.preventDefault();
     }
   }, []);
